+++ src/types/index.ts
@@ -11,6 +11,13 @@ import { IPrimaryUserParams, TFailLevelKey } from '../core/internalTypes'
 
 export type TJSObject = any // NOTE: Currently must be any! Not unknown or Record<string, unknown> or anything else, since linting etc will render this as error/unknow.
 
+// Human label types.
+export type TPersistThreshold =
+    | '0-Ignore-Errors' // 0 - Don't bail/fail on error, persist and try to recover.
+    | '1-Abort-on-Errors' // 1 - Stop parsing on the first error.
+    | '2-Abort-Even-on-Warnings' // 2 - Stop parsing on the first warning or error.
+export type TBailSensitivityLevel = 0 | 1 | 2 // Bail sensitivity level.
+
 export type TOnDuplicateKey =
     | 'warn-and-keep-first' // Keep last with a warning.
     | 'warn-and-overwrite' // 'warn-and-overwrite' = 'warn-and-keep-last'
@@ -24,6 +31,16 @@ export type TPreferredFailLevel = 'auto' | TFailLevelKey
 // | 'warnings-and-errors'
 
 // --- Public (user-facing) Interfaces -----------------------------------------------------------
+
+//{ line: 12, column: 8, type: 'Syntax-Error', message1: 'Invalid number' }
+export interface IIssuePayload {
+    line: number | undefined // NOTE: 1-based, so line 0 does not exist, set to undefined instead.
+    column: number | undefined // NOTE: 1-based, so column 0 does not exist, set to undefined instead.
+    typeKey: string // Transformed from the corresponding type, keep it lowercase since it's shown in meta, easier for tooling.
+    message: string
+    advice: string | undefined
+    hint: string | undefined
+}
 
 /**
  * @param failLevel - Minimum severity that should cause the parse to fail.
@@ -65,4 +82,80 @@ export interface IAllUserOptions extends IPrimaryUserParams {
     treatEmptyValueAsNull?: 'allow' | 'allow-with-warning' | 'disallow'
     onDuplicateKey?: TOnDuplicateKey
     // }
+}
+
+// --- Result Meta Interface (Public / User-facing)-------------------------
+
+/**
+ * @note NOTE: USE NO TYPES here (or neither in nested structures), they
+ *       should only be strings (and content MUST be transformed from the
+ *       types to lower snake case)
+ *
+ *       *** DUE TO make easier for tooling ***
+ *
+ * @note UPDATE 'metaSchemaVersion' on any edits to the meta structure.
+ */
+export interface IResultMetaData {
+    parserVersion: string
+    mode: 'lenient' | 'strict'
+    totalErrors: number
+    totalWarnings: number
+    totalMessages: number
+    runStartedAt: string
+    runFinishedAt: string
+    durationMs: number
+    preservesOrder: boolean // Member/section order: platform-, implementation-, and language-specific. Not mandated by the YINI spec.
+    orderGuarantee: 'implementation-defined' // De facto yes, in this specific implementation.
+    orderNotes?: string
+    source: {
+        sourceType: string // Transformed from the type, keep it lowercase since it's shown in resulted meta, easier for tooling.
+        fileName: undefined | string // Path and file name if from file.
+        hasDocumentTerminator: boolean
+        hasYiniMarker: boolean
+        byteSize: null | number
+        lineCount: null | number // Number of lines of content or in the yini file (before possible tampering).
+        sha256: null | string // SHA-256 hash of the original source content or file (before possible tampering).
+    }
+    structure: {
+        maxDepth: null | number
+        sectionCount: null | number // Section (header) count, section '(root)' NOT included.
+        memberCount: null | number // Section (member) count.
+        keysParsedCount: null | number // Including keys inside inline objects, and in section members.
+        // objectCount: null | number // (?) Incl. sections (objects) + inline objects.
+        // listCount: null | number
+        sectionNamePaths: string[] | null // All key/access paths to section Headers.
+    }
+    metaSchemaVersion: '1.1.0'
+    diagnostics?: {
+        failLevel: {
+            // preferredLevel: null | 'auto' | 0 | 1 | 2 // Input level into function.
+            preferredLevel: null | TPreferredFailLevel // Input level into function.
+            levelUsed: TBailSensitivityLevel
+            levelKey: TFailLevelKey // Mapped from the corresponding type, keep it lowercase since it's shown in meta, easier for tooling.
+            levelLabel: TPersistThreshold
+            levelDescription: string | null
+        }
+        errors: { errorCount: number; payload: IIssuePayload[] }
+        warnings: { warningCount: number; payload: IIssuePayload[] }
+        notices: { noticeCount: number; payload: IIssuePayload[] }
+        infos: { infoCount: number; payload: IIssuePayload[] }
+        environment: {
+            NODE_ENV: undefined | string
+            APP_ENV: undefined | string
+            lib: {
+                nodeEnv: undefined | string
+                appEnv: undefined | string
+                flags: { isDev: boolean; isDebug: boolean }
+            }
+        }
+        optionsUsed: IAllUserOptions
+    }
+    timing?: {
+        total: null | { timeMs: number; name: string }
+        phase0: undefined | { timeMs: number; name: string } // NOTE: When source is 'inline' phase0 is set to undefined as phase0 times I/O.
+        phase1: null | { timeMs: number; name: string }
+        phase2: null | { timeMs: number; name: string }
+        phase3: null | { timeMs: number; name: string }
+        phase4: null | { timeMs: number; name: string }
+    }
 }